# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AI4BMI API Groupe 3",
        "description": "API REST pour le système BMI — Gestion équipements et E-commerce BMI",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/v1/equipment-categories": {
            "get": {
                "tags": [
                    "Catégories Équipements"
                ],
                "summary": "Lister toutes les catégories",
                "description": "Retourne la liste des catégories. Le champ `equipments_count` permet de connaître le volume d'appareils par catégorie.",
                "operationId": "a5ba651c97b03393bb9ff4dc7936dd48",
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Maintenance Industrielle"
                                            },
                                            "equipments_count": {
                                                "description": "Nombre d'équipements liés",
                                                "type": "integer",
                                                "example": 12
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Session non authentifiée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Catégories Équipements"
                ],
                "summary": "Créer une nouvelle catégorie",
                "description": "Enregistre une catégorie. Utile pour filtrer les équipements par type (ex: Robotique, Informatique).",
                "operationId": "c7aa62c73a4c41bc4f7b78356fc8c747",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nom unique de la catégorie",
                                        "type": "string",
                                        "example": "Robotique"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Catégorie créée avec succès"
                    },
                    "403": {
                        "description": "Accès refusé — Droits admin requis"
                    },
                    "422": {
                        "description": "Données invalides (ex: nom déjà pris)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/equipment-categories/{id}": {
            "get": {
                "tags": [
                    "Catégories Équipements"
                ],
                "summary": "Voir le détail d'une catégorie",
                "description": "Retourne les informations de la catégorie et la liste des équipements associés.",
                "operationId": "6c4bd1e233a0bbaa2766b867288fd064",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID unique de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails retournés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "equipments": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Bras Articulé X-200"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Catégorie introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Catégories Équipements"
                ],
                "summary": "Modifier une catégorie",
                "description": "Permet de renommer une catégorie. La contrainte d'unicité du nom est conservée.",
                "operationId": "53f9d3f94df32c29781613adc003984a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Robotique Industrielle"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Catégorie renommée avec succès"
                    },
                    "403": {
                        "description": "Accès interdit"
                    },
                    "404": {
                        "description": "Catégorie introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Catégories Équipements"
                ],
                "summary": "Supprimer une catégorie",
                "description": "Action irréversible. Impossible si la catégorie contient encore des équipements.",
                "operationId": "3ea05e39f4ad3c003ca0268e64732133",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catégorie supprimée"
                    },
                    "403": {
                        "description": "Droits admin requis"
                    },
                    "409": {
                        "description": "Conflit : Catégorie non vide (équipements liés)"
                    },
                    "404": {
                        "description": "Catégorie introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)"
            }
        }
    },
...
}